feat: added Krux Editor project
• separated premake files
• fixed Camera.cpp warnings

+++ Krux/src/Krux/Render/Camera.cpp
@@ -4,14 +4,14 @@
 namespace Krux {
 
 	Camera::Camera(uint32_t width, uint32_t height)
-		: m_Width(width), m_Height(height), m_AspectRatio(m_Width / m_Height)
+		: m_Width((float)width), m_Height((float)height), m_AspectRatio(m_Width / m_Height)
 	{
 	}
 
 	void Camera::SetViewport(uint32_t width, uint32_t height)
 	{
-		m_Width = width;
-		m_Height = height;
+		m_Width = (float)width;
+		m_Height = (float)height;
 
 		m_AspectRatio = m_Width / m_Height;
 		UpdateProjection();
